ml-model-server/requirements.txt
fastapi==0.111.0
joblib==1.3.2
nltk==3.8.1
pandas==2.2.2
pydantic==2.8.0
scikit_learn==1.4.1.post1
uvicorn==0.30.1
python-dotenv==1.0.1